requirements.txt
streamlit>=1.30.0
openai>=1.12.0
pyodbc
requests>=2.31.0
python-dotenv>=1.0.0
googletrans==4.0.0-rc1
